// src/api/generated/model/listReviews200ItemsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Nero API
 * A Nero API é uma interface RESTful robusta projetada especificamente para o varejo de moda online. Ela oferece controle total sobre o ciclo de vida do produto, desde a entrada de estoque em múltiplas variantes (SKUs) até o processamento final do checkout. Ideal para marketplaces, lojas boutique ou aplicativos móveis de moda.
 * OpenAPI spec version: 1.0.0
 */
import type { ListReviews200ItemsItemLikes } from './listReviews200ItemsItemLikes';
import type { ListReviews200ItemsItemMediaItem } from './listReviews200ItemsItemMediaItem';
import type { ListReviews200ItemsItemUser } from './listReviews200ItemsItemUser';

export type ListReviews200ItemsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  rating: number;
  title: string | null;
  comment: string | null;
  variant: string | null;
  isVerified: boolean;
  createdAt: string;
  user: ListReviews200ItemsItemUser;
  likes: ListReviews200ItemsItemLikes;
  media: ListReviews200ItemsItemMediaItem[];
};